Corrects value for $entityIdName if Magento 2 Commerce is installed with Magento_CatalogStaging disabled
With appropriate changes to the XML file for the Map step, we can get past the Integrity Check step on a commerce-to-commerce or opensource-to-opensource migration, when having installed Magento Commerce with Magento_CatalogStaging disabled. However despite a successful integrity check, the Map step would fail while in Data mode due to a hard coded assumption that row_id would be present in the catalog_product_entity_media_gallery_value_to_entity table for any migration that is not opensource-to-opensource.

Error Resolved: SQLSTATE[42S22]: Column not found: 1054 Unknown column 'row_id' in 'field list'